requirements.txt
Flask==3.1.0
ifcopenshell==0.8.2
sentence-transformers==4.1.0
faiss-cpu==1.10.0
numpy==2.2.4
requests==2.32.3
Werkzeug==3.1.3 